requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
discord.py==2.4.0
requests==2.32.3
python-dotenv==1.0.1
